old_python_example/requirements.txt
aioboto3==15.5.0
aiohttp==3.13.3
aiofiles==25.1.0
fastapi==0.128.0
pydantic==2.12.5
uvicorn==0.40.0
tortoise-orm==0.25.3
python-multipart==0.0.21
